+++ gulpfile.js
@@ -11,6 +11,7 @@ var rename = require('gulp-rename');
 var tslint = require('gulp-tslint');
 var sassLint = require('gulp-sass-lint');
 var ts = require('gulp-typescript');
+var replace = require('gulp-replace');
 
 // PostCSS plugins
 var postcssSVG = require('postcss-svg');
@@ -29,6 +30,8 @@ var del = require('del');
 var runSequence = require('run-sequence');
 var browserSync = require('browser-sync');
 var merge = require('merge2');
+var nativeTslint = require('tslint');
+var npmPackage = require('./package.json');
 
 var paths = {
   source: {
@@ -61,9 +64,13 @@ gulp.task('clean', del.bind(null, [paths.target.html]));
 
 // TypeScript linting
 gulp.task('lint-ts', function() {
+  // The program is required for type checking rules to work: https://palantir.github.io/tslint/usage/type-checking/
+  var program = nativeTslint.Linter.createProgram("./tsconfig.json");
+
   return gulp.src(paths.source.ts)
   .pipe(tslint({
     formatter: 'verbose',
+    program: program,
   }))
   .pipe(tslint.report({
     // Print just the number of errors (instead of printing all errors again)
@@ -109,6 +116,7 @@ gulp.task('browserify', function() {
   // Compile output JS file
   var stream = browserifyBundle
   .pipe(source('bitmovinplayer-ui.js'))
+  .pipe(replace('{{VERSION}}', npmPackage.version))
   .pipe(buffer()) // required for production/sourcemaps
   .pipe(gulp.dest(paths.target.js));