+++ src/ts/main.ts
@@ -45,7 +45,22 @@ import {AirPlayToggleButton} from './components/airplaytogglebutton';
 import {VolumeSlider} from './components/volumeslider';
 import {PictureInPictureToggleButton} from './components/pictureinpicturetogglebutton';
 import {Spacer} from './components/spacer';
-import {ArrayUtils, StringUtils, PlayerUtils, UIUtils, BrowserUtils} from './utils';
+import {BackgroundColorSelectBox} from './components/subtitlesettings/backgroundcolorselectbox';
+import {BackgroundOpacitySelectBox} from './components/subtitlesettings/backgroundopacityselectbox';
+import {CharacterEdgeSelectBox} from './components/subtitlesettings/characteredgeselectbox';
+import {FontColorSelectBox} from './components/subtitlesettings/fontcolorselectbox';
+import {FontFamilySelectBox} from './components/subtitlesettings/fontfamilyselectbox';
+import {FontOpacitySelectBox} from './components/subtitlesettings/fontopacityselectbox';
+import {FontSizeSelectBox} from './components/subtitlesettings/fontsizeselectbox';
+import {SubtitleSettingsButton} from './components/subtitlesettings/subtitlesettingsbutton';
+import {SubtitleSettingsCloseButton} from './components/subtitlesettings/subtitlesettingsclosebutton';
+import {SubtitleSettingSelectBox} from './components/subtitlesettings/subtitlesettingselectbox';
+import {SubtitleSettingsLabel} from './components/subtitlesettings/subtitlesettingslabel';
+import {SubtitleSettingsOpenButton} from './components/subtitlesettings/subtitlesettingsopenbutton';
+import {SubtitleSettingsPanel} from './components/subtitlesettings/subtitlesettingspanel';
+import {WindowColorSelectBox} from './components/subtitlesettings/windowcolorselectbox';
+import {WindowOpacitySelectBox} from './components/subtitlesettings/windowopacityselectbox';
+import {ArrayUtils, StringUtils, PlayerUtils, UIUtils, BrowserUtils, ColorUtils, StorageUtils} from './utils';
 
 // Object.assign polyfill for ES5/IE9
 // https://developer.mozilla.org/de/docs/Web/JavaScript/Reference/Global_Objects/Object/assign
@@ -73,6 +88,7 @@ if (typeof Object.assign !== 'function') {
 
 // Expose classes to window
 (window as any).bitmovin.playerui = {
+  version: '{{VERSION}}',
   // Management
   UIManager,
   UIInstanceManager,
@@ -82,6 +98,8 @@ if (typeof Object.assign !== 'function') {
   PlayerUtils,
   UIUtils,
   BrowserUtils,
+  ColorUtils,
+  StorageUtils,
   // Components
   AdClickOverlay,
   AdMessageLabel,
@@ -131,4 +149,22 @@ if (typeof Object.assign !== 'function') {
   VolumeToggleButton,
   VRToggleButton,
   Watermark,
+  // Subtitle related components
+  subtitlesettings: {
+    BackgroundColorSelectBox,
+    BackgroundOpacitySelectBox,
+    CharacterEdgeSelectBox,
+    FontColorSelectBox,
+    FontFamilySelectBox,
+    FontOpacitySelectBox,
+    FontSizeSelectBox,
+    SubtitleSettingsButton,
+    SubtitleSettingsCloseButton,
+    SubtitleSettingSelectBox,
+    SubtitleSettingsLabel,
+    SubtitleSettingsOpenButton,
+    SubtitleSettingsPanel,
+    WindowColorSelectBox,
+    WindowOpacitySelectBox,
+  },
 };