+++ src/ts/utils.ts
@@ -348,3 +348,84 @@ export namespace BrowserUtils {
 
   export const isAndroid = navigator && navigator.userAgent && /Android/.test(navigator.userAgent);
 }
+
+export namespace ColorUtils {
+  export class Color {
+    r: number;
+    g: number;
+    b: number;
+    a: number;
+
+    constructor(r: number, g: number, b: number, a: number = 1) {
+      this.r = r;
+      this.g = g;
+      this.b = b;
+      this.a = a;
+    }
+
+    toCSS(): string {
+      return `rgba(${this.r}, ${this.g}, ${this.b}, ${this.a})`;
+    }
+  }
+
+  /**
+   * Parses a string formated as 'rgba(number, number, number, number)'
+   * and returns a color element of the same value
+   */
+  export function colorFromCss(css: string, fallback: Color = new Color(0, 0, 0)): Color {
+    if (!css.startsWith('rgba(')) {
+      return fallback;
+    }
+    let end = css.indexOf(')');
+    if (end !== css.length - 1) {
+      return fallback;
+    }
+    if (end !== css.lastIndexOf(')')) {
+      return fallback;
+    }
+    let vals = css.slice(5, end).split(',');
+    if (vals.length !== 4) {
+      return fallback;
+    }
+    return new Color(Number(vals[0]), Number(vals[1]), Number(vals[2]), Number(vals[3]));
+  }
+
+  // Transparent black and opaque white default colors for background and foreground
+  export const background = new Color(0, 0, 0, 0);
+  export const foreground = new Color(255, 255, 255, 1);
+}
+
+export namespace StorageUtils {
+  let hasLocalStorageCache: boolean;
+
+  export function hasLocalStorage(): boolean {
+    if (hasLocalStorageCache) {
+      return hasLocalStorageCache;
+    }
+
+    // hasLocalStorage is used to safely ensure we can use localStorage
+    // taken from https://developer.mozilla.org/en-US/docs/Web/API/Web_Storage_API/Using_the_Web_Storage_API#Feature-detecting_localStorage
+    let storage = window['localStorage'];
+    try {
+        let x = '__storage_test__';
+        storage.setItem(x, x);
+        storage.removeItem(x);
+        hasLocalStorageCache = true;
+    }
+    catch (e) {
+        hasLocalStorageCache = e instanceof DOMException && (
+            // everything except Firefox
+            e.code === 22 ||
+            // Firefox
+            e.code === 1014 ||
+            // test name field too, because code might not be present
+            // everything except Firefox
+            e.name === 'QuotaExceededError' ||
+            // Firefox
+            e.name === 'NS_ERROR_DOM_QUOTA_REACHED') &&
+            // acknowledge QuotaExceededError only if there's something already stored
+            storage.length !== 0;
+    }
+    return hasLocalStorageCache;
+  }
+}